+++ BasketApp/Product.cs
@@ -4,13 +4,13 @@ namespace BasketApp
 {
     public class Product
     {
-        public Product(string name, double cost)
+        public Product(string name, decimal cost)
         {
             this.Name = name;
             this.Cost = cost;
         }
 
         public string Name { get; set; }
-        public double Cost { get; set; }
+        public decimal Cost { get; set; }
     }
 }
